requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
install==1.3.5
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
urllib3==2.0.6
yarl==1.9.2
